VZ-10283 - Extract workloads and map to YAML objects
