+++ tools/oam-converter/pkg/constants/constants.go
@@ -1,0 +1,21 @@
+package constants
+
+const (
+	YamlComponents            = "components"
+	YamlSpec                  = "spec"
+	YamlTraits                = "traits"
+	TraitKind                 = "kind"
+	TraitComponent            = "trait"
+	YamlMetadata              = "metadata"
+	YamlName                  = "name"
+	YamlNamespace             = "namespace"
+	IngressTrait              = "IngressTrait"
+	MetricsTrait              = "metricsTrait"
+	GatewayAPIVersion         = "networking.istio.io/v1alpha3"
+	VirtualServiceAPIVersion  = "networking.istio.io/v1alpha3"
+	DestinationRuleAPIVersion = "networking.istio.io/v1alpha3"
+	AuthorizationAPIVersion   = "security.istio.io/v1beta1"
+	HttpsProtocol             = "HTTPS"
+	VerrazzanoClusterIssuer   = "verrazzano-cluster-issuer"
+	CertificateAPIVersion     = "cert-manager.io/v1"
+)
