+++ tools/oam-converter/vs.yaml
@@ -1,0 +1,68 @@
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  creationTimestamp: null
+  name: bobs-books-rule-0-vs
+  namespace: bobs-books
+spec:
+  gateways:
+  - bobs-books-bobs-books-gw
+  hosts:
+  - bobs-books.bobs-books.172.18.0.230.nip.io
+  http:
+  - match:
+    - uri:
+        prefix: /bobs-bookstore-order-manager/orders
+    - uri:
+        prefix: /bobs-bookstore-order-manager/books
+    route:
+    - destination:
+        host: bobs-bookstore-cluster-cluster-1
+        port:
+          number: 8001
+status: {}
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  creationTimestamp: null
+  name: bobs-books-rule-0-vs
+  namespace: bobs-books
+spec:
+  gateways:
+  - bobs-books-bobs-books-gw
+  hosts:
+  - bobs-books.bobs-books.172.18.0.230.nip.io
+  http:
+  - match:
+    - uri:
+        prefix: /
+    route:
+    - destination:
+        host: robert-helidon
+        port:
+          number: 8080
+status: {}
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  creationTimestamp: null
+  name: bobs-books-rule-0-vs
+  namespace: bobs-books
+spec:
+  gateways:
+  - bobs-books-bobs-books-gw
+  hosts:
+  - bobs-books.bobs-books.172.18.0.230.nip.io
+  http:
+  - match:
+    - uri:
+        prefix: /bobbys-front-end
+    route:
+    - destination:
+        host: bobbys-front-end-cluster-cluster-1
+        port:
+          number: 8001
+status: {}
+---
